+++ libsel4vmmplatsupport/src/ioports.c
@@ -18,7 +18,7 @@
 
 static int io_port_compare_by_range(const void *pkey, const void *pelem)
 {
-    unsigned int key = (unsigned int)pkey;
+    unsigned int key = (unsigned int)(uintptr_t)pkey;
     const ioport_entry_t *entry = (const ioport_entry_t *)(*(const ioport_entry_t **)pelem);
     const ioport_range_t *elem = &entry->range;
     if (key < elem->start) {
